handle error on delete author

+++ src/main/java/com/example/demo/service/impl/AuthorServiceImpl.java
@@ -2,10 +2,14 @@ package com.example.demo.service.impl;
 
 import com.example.demo.dto.AuthorDto;
 import com.example.demo.entities.Author;
+import com.example.demo.error.BadRequestError;
 import com.example.demo.repository.AuthorRepository;
 import com.example.demo.modelmapper.ModelMapperUtil;
+import com.example.demo.repository.BookRepository;
 import com.example.demo.service.AuthorService;
 import org.modelmapper.ModelMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,7 +23,15 @@ public class AuthorServiceImpl implements AuthorService {
     private AuthorRepository authorRepository;
 
     @Autowired
+    private BookRepository bookRepository;
+
+    @Autowired
     private ModelMapper modelMapper;
+
+    Logger logger = LoggerFactory.getLogger(AuthorServiceImpl.class);
+
+    static final String ITEM_TYPE = "author";
+    static final String WITH_ID_NOT_FOUND = "{} with id {} not found";
 
     @Override
     public List<AuthorDto> findAll() {
@@ -40,7 +52,16 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public boolean delete(long id) {
-        //TODO
+        if (!authorRepository.existsById(id)) {
+            logger.warn(WITH_ID_NOT_FOUND, ITEM_TYPE);
+            return false;
+        }
+
+        var numberOfBookLinked = bookRepository.countByAuthorId(id);
+        if (numberOfBookLinked > 0) {
+            throw BadRequestError.deletionNotAllowedLinkedItem(ITEM_TYPE, "book", numberOfBookLinked);
+        }
+
         return authorRepository.findById(id)
                 .map(authorEntity -> {
                     authorRepository.delete(authorEntity);
